+++ test/sample/test2.in.dart
@@ -1,0 +1,8 @@
+library params.test;
+
+import 'package:params/annotation.dart';
+
+class SimpleTest2 extends Object with Serialize {
+  @ModelParameter(name:"other_field1_name")
+  String field1;
+}