+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+- rename Converter -> Accessor
+- remove method asJsObject
+- rename method asMap -> toJson
+- support json output name.
+- support Accessor
+
 ## 0.1.0
 
 - Initial